// src/services/api/generated/inventory-service/model/createRoomRequestRoomType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */

export type CreateRoomRequestRoomType = typeof CreateRoomRequestRoomType[keyof typeof CreateRoomRequestRoomType];


export const CreateRoomRequestRoomType = {
  STANDARD: 'STANDARD',
  IMAX: 'IMAX',
  FOUR_DX: 'FOUR_DX',
  SCREEN_X: 'SCREEN_X',
  VIP: 'VIP',
} as const;