// src/services/api/generated/inventory-service/model/createRoomRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */
import type { CreateRoomRequestRoomType } from './createRoomRequestRoomType';

export interface CreateRoomRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  roomType: CreateRoomRequestRoomType;
}