// src/services/api/generated/inventory-service/model/roomResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */
import type { RoomResponseRoomType } from './roomResponseRoomType';

export interface RoomResponse {
  id?: string;
  cinemaId?: string;
  name?: string;
  roomType?: RoomResponseRoomType;
  active?: boolean;
  createdAt?: string;
  updatedAt?: string;
}